Updated DuplicateKeyFrame and SplitKeyFrame
Updated DuplicateKeyFrame and SplitKeyFrame to account for grouped strokes. Previously it did not recognize grouped strokes, so when it was replicated to the new canvas, there was no key to the strokes, preventing users from selecting the stroke. When un-registered grouped strokes was selected, it will select the original grouped strokes in the previous canvas.

Refactored blocks of code in DuplicateKeyFrame and SplitKeyFrame onto ReplicateStrokes.