Fixes on Adding/Removing Layers and SplitKeyFrames
Fixed line of code in SketchWriter where it did not have an if-statement on the local dictionary. This resolves the issue with "deleted" canvases in the animation
grid. Also fixed bug where deleted tracks are showing up when playing the timeline. "Deleted" canvases will not be saved as the SketchWriter will only recognize
non-deleted canvases.

Fixed SplitKeyFrames so it will return early if the head of the frame is selected.