Fixes on ReplicateStrokes and RemoveKeyFrame. QOL update.
Fixed ReplicateStrokes to account for Batches that do not have Geometry. This occurs when "erasing" joined strokes. The batch will
be memorized, but there is no geometry. If no geometry, then we'll uncreate the batch.

Fixed RemoveKeyFrame to break up a keyframe that has been extended. An empty extended keyframe will return an out-of-index error.

updateFrameInfo is changed to disambiguate the UI. A colon implies there are seconds and nano-seconds when it's not the case.